1400-25

+++ 1400/C_Basketball_Exercise.cpp
@@ -2,12 +2,12 @@
 #include<vector>
 using namespace std;
 
-int solve(vector<int> &a, vector<int> &b, int i, bool turn, int n, vector<vector<int>> &dp){
+long long solve(vector<int> &a, vector<int> &b, int i, bool turn, int n, vector<vector<long long>> &dp){
     if(i == n) return 0;
     if(dp[i][turn] != -1) return dp[i][turn];
 
-    int take = solve(a, b, i + 1, !turn, n, dp) + (turn ? a[i] : b[i]);
-    int leave = solve(a, b, i + 1, turn, n, dp);
+    long long take = solve(a, b, i + 1, !turn, n, dp) + (turn ? a[i] : b[i]);
+    long long leave = solve(a, b, i + 1, turn, n, dp);
 
     return dp[i][turn] = max(take, leave);
 
@@ -24,8 +24,8 @@ int main(){
     for(int i = 0; i < n; i++){
         cin >> b[i];
     }
-    vector<vector<int>> dp(n, vector<int>(2, -1));
-    int res = solve(a, b, 0, true, n, dp);
+    vector<vector<long long>> dp(n, vector<long long>(2, -1));
+    long long res = solve(a, b, 0, true, n, dp);
     res = max(res, solve(a, b, 0, false, n, dp));
 
     cout << res << endl;
